Fixup broken case submission (#146)
* convert notifications methods to noops

* fix and refactor

The structure of cases.go was/is a mess. Hopefully the logic is a bit cleaner now. The behaviour of the api is also more obvious, with POST requests returning JSON validation objects when errors are encountered. The frontend still needs to be updated to handle this data appropriately in the way that the `individual.gohtml` template currently does. As an aside, I believe it might be worthwhile to move some of the validation logic to the client side, at least as far as syntax checking is concerned (emails, phone numbers, etc). Given that Core is intended to function in environments with spotty connection, we should reserve server-side validation for evaluations which are impossible to do on the client host.

* Update cases.go

* update cases.go

* refactor

* fix failing test

* fix failing test; update e2e cli options